gui: ncurses: use size_t for index subscript
Fix warning:

    ../../modules/gui/ncurses.c: In function ‘DrawStats’:
    ../../modules/gui/ncurses.c:691:23: warning: comparison of integer expressions of different signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
      691 |     for (int i = 0; i < item->es_vec.size ; i++) {
          |                       ^

+++ modules/gui/ncurses.c
@@ -688,7 +688,7 @@ static int DrawStats(intf_thread_t *intf)
     vlc_mutex_lock(&item->lock);
     p_stats = item->p_stats;
 
-    for (int i = 0; i < item->es_vec.size ; i++) {
+    for (size_t i = 0; i < item->es_vec.size ; i++) {
         i_audio += (item->es_vec.data[i].es.i_cat == AUDIO_ES);
         i_video += (item->es_vec.data[i].es.i_cat == VIDEO_ES);
     }
